refactor file and resource folder structure
