+++ arcgis-java-toolkit-examples/src/main/java/com/esri/arcgisruntime/toolkit/examples/ExamplesApp.java
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-package com.esri.arcgisruntime.toolkit;
+package com.esri.arcgisruntime.toolkit.examples;
 
-import com.esri.arcgisruntime.toolkit.controller.ExamplesAppController;
+import com.esri.arcgisruntime.toolkit.examples.controller.ExamplesAppController;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.geometry.Rectangle2D;
@@ -37,7 +37,7 @@ public class ExamplesApp extends Application {
     @Override
     public void start(Stage primaryStage) {
         // loads the FXML
-        FXMLLoader loader = new FXMLLoader(getClass().getResource("/fxml/app.fxml"));
+        FXMLLoader loader = new FXMLLoader(getClass().getResource("app.fxml"));
         try {
             Parent root = loader.load();
             controller = loader.getController();
@@ -51,9 +51,9 @@ public class ExamplesApp extends Application {
             // configures the scene and sets it to the stage
             var scene = new Scene(root);
             // individual stylesheets can be commented out for testing purposes
-            scene.getStylesheets().add("/styles/style.css");
-            scene.getStylesheets().add("/styles/app.css");
-            scene.getStylesheets().add("/styles/example.css");
+            scene.getStylesheets().add(String.valueOf(getClass().getResource("styles/style.css")));
+            scene.getStylesheets().add(String.valueOf(getClass().getResource("styles/app.css")));
+            scene.getStylesheets().add(String.valueOf(getClass().getResource("styles/example.css")));
             primaryStage.setScene(scene);
             primaryStage.show();
         } catch (Exception e) {
