+++ arcgis-java-toolkit-examples/src/main/java/com/esri/arcgisruntime/toolkit/examples/utils/ExampleUtils.java
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-package com.esri.arcgisruntime.toolkit.utils;
+package com.esri.arcgisruntime.toolkit.examples.utils;
 
 import com.esri.arcgisruntime.ArcGISRuntimeEnvironment;
-import com.esri.arcgisruntime.toolkit.controller.ExampleView;
-import com.esri.arcgisruntime.toolkit.model.Example;
+import com.esri.arcgisruntime.toolkit.examples.controller.ExampleView;
+import com.esri.arcgisruntime.toolkit.examples.model.Example;
 import javafx.geometry.Pos;
 import javafx.geometry.Rectangle2D;
 import javafx.scene.Node;
@@ -37,7 +37,7 @@ import java.util.logging.Logger;
 
 /**
  * Defines util methods to support the configuration of Examples that are displayed within an
- * {@link com.esri.arcgisruntime.toolkit.controller.ExampleView}.
+ * {@link ExampleView}.
  *
  * @since 100.15.0
  */
@@ -110,8 +110,8 @@ public class ExampleUtils {
         // configure the Scene and add to the Stage
         var scene = new Scene(stackPane);
         // individual stylesheets can be commented out for testing purposes
-        scene.getStylesheets().add("/styles/example.css");
-        scene.getStylesheets().add("/styles/style.css");
+        scene.getStylesheets().add("/com/esri/arcgisruntime/toolkit/examples/styles/example.css");
+        scene.getStylesheets().add("/com/esri/arcgisruntime/toolkit/examples/styles/style.css");
         primaryStage.setScene(scene);
         primaryStage.show();
     }
