+++ arcgis-java-toolkit-examples/src/main/resources/com/esri/arcgisruntime/toolkit/examples/controller/images/Compass.png
