+++ arcgis-java-toolkit/src/main/resources/com/esri/arcgisruntime/toolkit/feature-template-picker.css
