+++ arcgis-java-toolkit/src/main/resources/com/esri/arcgisruntime/toolkit/floor-filter.css
