+++ arcgis-java-toolkit/src/main/resources/com/esri/arcgisruntime/toolkit/skins/utility_network_trace.fxml
